cleanup added more methods
